moving controllers to new src directory and resetting system as needed
