+++ public/index.php
@@ -3,12 +3,12 @@
 session_start();
 
 $config = require_once('../config.php');
-require_once '../MasterController.php';
+require_once '../src/MasterController.php';
 
-require_once '../Comment.php';
-require_once '../User.php';
-require_once '../Story.php';
-require_once '../Index.php';
+require_once '../src/Comment.php';
+require_once '../src/User.php';
+require_once '../src/Story.php';
+require_once '../src/Index.php';
 
 $framework = new MasterController($config);
 echo $framework->execute();
