+++ src/Index.php
@@ -36,7 +36,7 @@ class Index {
         
         $content .= '</ol>';
         
-        require 'layout.phtml';
+        require '../layout.phtml';
     }
 }
